+++ index.html
@@ -10,6 +10,6 @@
     <div id="root">
         <h1>Not Rendered</h1>
     </div>
-    <script type="module" src="App.js"></script>
+    <script type="module" src="./src/App.js"></script>
 </body>
 </html>
