+++ src/App.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import TilesComp from './components/Tiles';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(<TilesComp />);
