+++ src/assets/data/tiles.json
@@ -1,0 +1,24 @@
+{
+    "data" : [
+        {
+            "id": 1,
+            "tileName": "Roasted Chicken",
+            "rating": "4"
+        },
+        {
+            "id": 2,
+            "tileName": "Salad",
+            "rating": "3"
+        },
+        {
+            "id": 3,
+            "tileName": "Chinese",
+            "rating": "2"
+        },
+        {
+            "id": 4,
+            "tileName": "Punjabi",
+            "rating": "1"
+        }
+    ]
+}
